services/gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
prometheus-client==0.20.0
python-dotenv==1.0.1
